@@ -11,42 +11,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "core-js": "^2.5.4",
-    "ngx-base": "^3.3.1",
-    "ngx-login-client": "^2.5.1",
-    "patternfly": "^3.54.8",
+    "@angular/animations": "^7.1.3",
+    "@angular/common": "^7.1.3",
+    "@angular/compiler": "^7.1.3",
+    "@angular/core": "^7.1.3",
+    "@angular/forms": "^7.1.3",
+    "@angular/http": "^7.1.3",
+    "@angular/platform-browser": "^7.1.3",
+    "@angular/platform-browser-dynamic": "^7.1.3",
+    "@angular/router": "^7.1.3",
+    "core-js": "^2.6.0",
+    "ngx-base": "^4.0.0",
+    "ngx-login-client": "^2.8.0",
+    "ngx-widgets": "^3.2.0",
+    "patternfly": "^3.58.0",
     "patternfly-ng": "^4.5.1",
-    "rxjs": "~6.2.0",
+    "rxjs": "~6.3.3",
+    "tslib": "^1.9.0",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.0",
-    "@angular/cli": "~6.2.3",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@types/jasmine": "^2.8.9",
-    "@types/jasminewd2": "^2.0.5",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "^7.1.2",
+    "@angular/compiler-cli": "^7.1.3",
+    "@angular/language-service": "^7.1.3",
+    "@types/jasmine": "^3.3.1",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/node": "~10.12.12",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~3.1.3",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "typescript": "^3.1.1"
   }
 }
