@@ -1,7 +1,8 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { UserService, User } from 'ngx-login-client';
+import { User } from 'ngx-login-client';
 import { Subscription } from 'rxjs';
 import { UserStore } from '../../store/user.store';
+import { UserService } from 'src/app/services/user.service';
 
 @Component({
   selector: 'app-users-container',
@@ -9,32 +10,25 @@ import { UserStore } from '../../store/user.store';
   styleUrls: ['./users-container.component.css']
 })
 export class UsersContainerComponent implements OnInit, OnDestroy {
-
   private subscriptions: Subscription = new Subscription();
   users: User[];
   isSubscriptionError: boolean;
   isSearchComplete: boolean;
 
-  constructor(
-    private userService: UserService,
-    private userStore: UserStore
-   ) { }
+  constructor(private userStore: UserStore, private userService: UserService) {}
 
-  ngOnInit() {
-    // this.isSearchComplete = false;
-  }
+  ngOnInit() {}
 
   searchUsers(searchTerm: string): void {
     this.isSearchComplete = false;
     this.subscriptions.add(
-      this.userService
-        .getUsersBySearchString(searchTerm)
-        .subscribe((users: User[]) => {
+      this.userService.getUsersByName(searchTerm).subscribe(
+        (users: User[]) => {
           this.users = users;
           this.isSearchComplete = true;
           this.userStore.addUsers(users);
         },
-         err => {
+        () => {
           this.isSubscriptionError = false;
         }
       )
