@@ -3,7 +3,11 @@ import { TestBed } from '@angular/core/testing';
 import { RequestCache } from './request-cache.service';
 
 describe('RequestCache', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  beforeEach(() => TestBed.configureTestingModule({
+    providers: [
+      RequestCache
+    ]
+  }));
 
   it('should be created', () => {
     const service: RequestCache = TestBed.get(RequestCache);
