@@ -1,3 +1,7 @@
 export const environment = {
-  production: true
+  production: true,
+  adminApiUrl: 'https://admin-console.api.openshift.io/api/',
+  authApiUrl: 'https://auth.prod.openshift.io/api/',
+  ssoApiUrl: 'https://sso.prod.openshift.io/api/',
+  witApiUrl: 'https://prod.openshift.io/api/'
 };
